Transferring logic from the UserDAO class to the UserService class and adding the UserValidator class and interface

+++ src/service/UserServiceImpl.java
@@ -6,6 +6,8 @@ import dao.UserDaoImpl;
 
 import enity.User;
 
+import exceptions.EntityNotFoundException;
+
 import exceptions.ObjectInitializeException;
 
 import exceptions.UserServiceException;
@@ -20,32 +22,63 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> readAllUsers() {
-        return userDao.readAllUsers();
+    public List<User> readAllUsers() throws UserServiceException {
+        try {
+            return userDao.readAllUsers();
+        } catch (ObjectInitializeException e) {
+            throw new UserServiceException(e);
+        }
     }
 
     @Override
     public void add(User user) throws UserServiceException {
         try {
-            userDao.add(user);
-        } catch (ObjectInitializeException e) {
+            if (user != null) {
+                userDao.add(user);
+            } else {
+                throw new EntityNotFoundException();
+            }
+        } catch (EntityNotFoundException e) {
             throw new UserServiceException(e);
         }
-
     }
 
     @Override
-    public User upDate(User user) {
-        return userDao.upDate(user);
+    public User upDate(User user, int index) throws UserServiceException {
+        try {
+            if (userDao.upDate(user, index) != null) {
+                return userDao.upDate(user, index);
+            } else {
+                throw new EntityNotFoundException();
+            }
+        } catch (EntityNotFoundException e) {
+            throw new UserServiceException(e);
+        }
     }
 
     @Override
     public User read(int userID) throws UserServiceException {
-        return userDao.read(userID);
+        try {
+            if (userDao.read(userID) != null) {
+                return userDao.read(userID);
+            } else {
+                throw new EntityNotFoundException();
+            }
+        } catch (EntityNotFoundException e) {
+            throw new UserServiceException(e);
+        }
     }
 
     @Override
     public void delete(User user) throws UserServiceException {
-        userDao.delete(user);
+        try {
+            if (user != null) {
+                userDao.delete(user);
+            } else {
+                throw new EntityNotFoundException("Удаляемый Вами пользователь не найден");
+            }
+        } catch (EntityNotFoundException e) {
+            throw new UserServiceException(e);
+        }
     }
 }
